Port UNIX V7 to Epoch68

+++ unix/v7/usr/include/varargs.h
@@ -1,5 +1,7 @@
+#include "sys/inttypes.h"
+
 typedef char *va_list;
-# define va_dcl int va_alist;
+# define va_dcl i16 va_alist;
 # define va_start(list) list = (char *) &va_alist
 # define va_end(list)
 # define va_arg(list,mode) ((mode *)(list += sizeof(mode)))[-1]
